Move components into separate files
